[fastlane] Seperate screenshot code from testing code
Fixed bugs regarding change of localization during screenshoting
